+++ regression-test/data/ddl_p0/test_create_view_nereids.out
@@ -202,7 +202,7 @@ test_view_from_view	CREATE VIEW `test_view_from_view` AS select `internal`.`regr
 7	1
 
 -- !test_backquote_in_view_define_sql --
-test_backquote_in_view_define	CREATE VIEW `test_backquote_in_view_define` AS select `internal`.`regression_test_ddl_p0`.`mal_test_view`.`a` AS `ab``c`, `internal`.`regression_test_ddl_p0`.`mal_test_view`.`b` AS `c2` from `internal`.`regression_test_ddl_p0`.`mal_test_view`;	utf8mb4	utf8mb4_0900_bin
+test_backquote_in_view_define	CREATE VIEW `test_backquote_in_view_define` AS select `internal`.`regression_test_ddl_p0`.`mal_test_view`.`a` AS `abc`, `internal`.`regression_test_ddl_p0`.`mal_test_view`.`b` AS `c2` from `internal`.`regression_test_ddl_p0`.`mal_test_view`;	utf8mb4	utf8mb4_0900_bin
 
 -- !test_backquote_in_table_alias --
 \N	6
@@ -225,6 +225,28 @@ test_backquote_in_view_define	CREATE VIEW `test_backquote_in_view_define` AS sel
 
 -- !test_backquote_in_table_alias_sql --
 test_backquote_in_table_alias	CREATE VIEW `test_backquote_in_table_alias` AS select `internal`.`regression_test_ddl_p0`.`ab``c`.`a` AS `c1`, `internal`.`regression_test_ddl_p0`.`ab``c`.`b` AS `c2` from (select `internal`.`regression_test_ddl_p0`.`mal_test_view`.`a`,`internal`.`regression_test_ddl_p0`.`mal_test_view`.`b` from `internal`.`regression_test_ddl_p0`.`mal_test_view`) `ab``c`;	utf8mb4	utf8mb4_0900_bin
+
+-- !test_invalid_column_name_in_table --
+\N	6
+1	2
+1	3
+1	4
+1	4
+1	7
+2	8
+3	5
+3	6
+3	9
+4	2
+5	\N
+5	6
+5	6
+5	6
+5	8
+7	1
+
+-- !test_invalid_column_name_in_table_define_sql --
+test_invalid_column_name_in_table	CREATE VIEW `test_invalid_column_name_in_table` AS select `internal`.`regression_test_ddl_p0`.`mal_test_view`.`a` ,`internal`.`regression_test_ddl_p0`.`mal_test_view`.`b` from `internal`.`regression_test_ddl_p0`.`mal_test_view`;	utf8mb4	utf8mb4_0900_bin
 
 -- !test_generate --
 1	10	A	30
